move hooks in lua folder to Misc.cpp (#234)
