+++ FTL_HyperspaceGCC.cbp
@@ -1319,13 +1319,7 @@
 		<Unit filename="libzhl.h">
 			<Option virtualFolder="ZHL/" />
 		</Unit>
-		<Unit filename="lua/InternalEvents.cpp">
-			<Option virtualFolder="Scripting/" />
-		</Unit>
 		<Unit filename="lua/InternalEvents.h">
-			<Option virtualFolder="Scripting/" />
-		</Unit>
-		<Unit filename="lua/LuaLibScript.cpp">
 			<Option virtualFolder="Scripting/" />
 		</Unit>
 		<Unit filename="lua/LuaLibScript.h">
@@ -1335,9 +1329,6 @@
 			<Option virtualFolder="Scripting/" />
 		</Unit>
 		<Unit filename="lua/LuaScriptInit.h">
-			<Option virtualFolder="Scripting/" />
-		</Unit>
-		<Unit filename="lua/RenderEvents.cpp">
 			<Option virtualFolder="Scripting/" />
 		</Unit>
 		<Unit filename="lua/RenderEvents.h">
